test(casedb): REFAC rename files, also has a fix to remove double line after the merge.
