+++ test/casedb/integration/edge_cases_access/conftest.py
@@ -1,26 +1,26 @@
 # These fixtures are used across multiple test modules in the integration test suite, so they are defined in conftest.py
 # to avoid duplication and ensure consistent setup of test users, organizations, and reference data for all tests.
 
-from test.casedb.integration.edge_cases_access.setup.setup_case_data import (  # noqa: F401
-    setup_case_data,
+from test.casedb.integration.edge_cases_access.setup.setup_case_data_operational import (  # noqa: F401
+    setup_case_data_operational,
 )
-from test.casedb.integration.edge_cases_access.setup.setup_case_type_data import (  # noqa: F401
-    setup_case_type_data,
+from test.casedb.integration.edge_cases_access.setup.setup_case_data_reference import (  # noqa: F401
+    setup_case_data_reference,
 )
-from test.casedb.integration.edge_cases_access.setup.setup_test_users_and_organizations import (  # noqa: F401
-    setup_test_users_and_organizations,
+from test.casedb.integration.edge_cases_access.setup.setup_test_users_and_organizations_operational import (  # noqa: F401
+    setup_test_users_and_organizations_operational,
 )
-from test.casedb.integration.edge_cases_access.setup.setup_test_users_and_organizations_op import (  # noqa: F401
-    setup_test_users_and_organizations_op,
+from test.casedb.integration.edge_cases_access.setup.setup_test_users_and_organizations_reference import (  # noqa: F401
+    setup_test_users_and_organizations_reference,
 )
 
 # Note: prevents formatters and linters from removing the imports as unused since they are used implicitly as fixtures in the test modules. The actual test modules should import the specific fixtures they use directly from this conftest.py for clarity,
 # but this dummy variable ensures that the imports are retained here.
 dummy = (
-    setup_test_users_and_organizations,
-    setup_test_users_and_organizations_op,
-    setup_case_type_data,
-    setup_case_data,
+    setup_test_users_and_organizations_reference,
+    setup_test_users_and_organizations_operational,
+    setup_case_data_reference,
+    setup_case_data_operational,
 )  # noqa: F841
