+++ test/casedb/integration/edge_cases_access/setup/setup_test_users_and_organizations_reference.py
@@ -1,5 +1,7 @@
 from test.casedb.casedb_test_client import CasedbTestClient as Env
-from test.casedb.integration.edge_cases_access.setup.define_edge_cases import EDGE_CASES
+from test.casedb.integration.edge_cases_access.setup.define_edge_cases_reference import (
+    EDGE_CASES,
+)
 
 import pytest
 
@@ -9,7 +11,7 @@ VERBOSE = False  # Set to True to enable detailed print statements during setup 
 
 
 @pytest.fixture(scope="module")
-def setup_test_users_and_organizations(env: Env) -> None:
+def setup_test_users_and_organizations_reference(env: Env) -> None:
     """
     Set up common test users and organizations driven by EDGE_CASES.
